requirements.txt
fastapi
uvicorn[standard]
pytest
sqlalchemy
httpx[socks]
pyyaml
pydantic
openai-whisper
memoryscope
python-multipart